plan-b-gcloud-with-vs-code/02-langchain-agent-projects/02_prompt_template.py
from dotenv import load_dotenv
load_dotenv()
from langchain_core.prompts import ChatPromptTemplate
prompt = ChatPromptTemplate.from_template("Tell a joke about {topic}")
print(prompt.format(topic="robots"))
