requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-tavily>=0.1.0
langgraph>=0.0.30
tavily-python>=0.3.0
python-dotenv>=1.0.0
langchain-ollama
langchain-google-genai
pytest>=8.0.0
